move class to avoid import loop
